feat: create cover creator types
